requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
pymongo
certifi
pymongo
pymongo[srv]==3.10
pyaml
mlflow

#-e .
